chore: refactor document transition to remove code duplication
