+++ lib/document/stateTransition/documentTransition/AbstractDocumentTransition.js
@@ -1,9 +1,13 @@
+const lodashGet = require('lodash.get');
+const lodashSet = require('lodash.set');
+
 /**
  * @abstract
  */
 class AbstractDocumentTransition {
-  constructor(rawDocumentTransition) {
+  constructor(rawDocumentTransition, dataContract) {
     this.dataContractId = rawDocumentTransition.$dataContractId;
+    this.dataContract = dataContract;
   }
 
   /**
@@ -23,15 +27,112 @@ class AbstractDocumentTransition {
   }
 
   /**
+   * Get data
+   *
+   * @return {Object}
+   */
+  getData() {
+    return {};
+  }
+
+  /**
+   * @abstract
+   *
+   * @return {Object}
+   */
+  getSystemProperties() {
+    throw new Error('Not implemented');
+  }
+
+  /**
    * Get JSON representation
    *
-   * @returns { { $action: string, $dataContractId: string } }
+   * @return {Object}
    */
   toJSON() {
+    const data = JSON.parse(
+      JSON.stringify(this.data),
+    );
+
+    const encodedProperties = this.dataContract.getEncodedProperties(
+      this.getType(),
+    );
+
+    Object.keys(encodedProperties)
+      .forEach((propertyPath) => {
+        const property = encodedProperties[propertyPath];
+
+        if (property.contentEncoding === 'binary') {
+          const value = lodashGet(data, propertyPath);
+          if (value !== undefined) {
+            lodashSet(
+              data,
+              propertyPath,
+              Buffer.from(value)
+                .toString('utf-8'),
+            );
+          }
+        }
+      });
+
     return {
       $action: this.getAction(),
       $dataContractId: this.getDataContractId(),
+      ...this.getSystemProperties(),
+      ...data,
     };
+  }
+
+  /**
+   * Get plain object representation
+   *
+   * @returns {Object}
+   */
+  toObject() {
+    return {
+      $action: this.getAction(),
+      $dataContractId: this.getDataContractId(),
+      ...this.getSystemProperties(),
+      ...this.data,
+    };
+  }
+
+  /**
+   * Create document transition from JSON
+   *
+   * @param {Class} documentTransitionClass
+   * @param {
+   *   RawDocumentCreateTransition | RawDocumentReplaceTransition | RawDocumentDeleteTransition
+   * } rawDocumentTransition
+   * @param {DataContract} dataContract
+   *
+   * @return {
+   *   RawDocumentCreateTransition | RawDocumentReplaceTransition | RawDocumentDeleteTransition
+   * }
+   */
+  static fromJson(documentTransitionClass, rawDocumentTransition, dataContract) {
+    const encodedProperties = dataContract.getEncodedProperties(
+      rawDocumentTransition.$type,
+    );
+
+    Object.keys(encodedProperties)
+      .forEach((propertyPath) => {
+        const property = encodedProperties[propertyPath];
+
+        if (property.contentEncoding === 'binary') {
+          const value = lodashGet(rawDocumentTransition, propertyPath);
+          if (value !== undefined) {
+            lodashSet(
+              rawDocumentTransition,
+              propertyPath,
+              Buffer.from(value, 'utf-8'),
+            );
+          }
+        }
+      });
+
+    // eslint-disable-next-line new-cap
+    return new documentTransitionClass(rawDocumentTransition, dataContract);
   }
 }
