+++ lib/document/stateTransition/documentTransition/DocumentCreateTransition.js
@@ -1,6 +1,3 @@
-const lodashGet = require('lodash.get');
-const lodashSet = require('lodash.set');
-
 const AbstractDocumentTransition = require('./AbstractDocumentTransition');
 
 class DocumentCreateTransition extends AbstractDocumentTransition {
@@ -9,9 +6,7 @@ class DocumentCreateTransition extends AbstractDocumentTransition {
    * @param {DataContract} dataContract
    */
   constructor(rawTransition, dataContract) {
-    super(rawTransition);
-
-    this.dataContract = dataContract;
+    super(rawTransition, dataContract);
 
     const data = { ...rawTransition };
 
@@ -102,64 +97,12 @@ class DocumentCreateTransition extends AbstractDocumentTransition {
   }
 
   /**
-   * Get document transition as a plain object
-   * with encoded binary fields
-   *
-   * @return {RawDocumentCreateTransition}
-   */
-  toJSON() {
-    const data = JSON.parse(
-      JSON.stringify(this.data),
-    );
-
-    const encodedProperties = this.dataContract.getEncodedProperties(
-      this.getType(),
-    );
-
-    Object.keys(encodedProperties)
-      .forEach((propertyPath) => {
-        const property = encodedProperties[propertyPath];
-
-        if (property.contentEncoding === 'binary') {
-          const value = lodashGet(data, propertyPath);
-          if (value !== undefined) {
-            lodashSet(
-              data,
-              propertyPath,
-              Buffer.from(value)
-                .toString('utf-8'),
-            );
-          }
-        }
-      });
-
-    return {
-      ...this.getSystemProperties(),
-      ...data,
-    };
-  }
-
-  /**
-   * Get document transition as a plain object
-   * without encoded binary fields
-   *
-   * @return {RawDocumentCreateTransition}
-   */
-  toObject() {
-    return {
-      ...this.getSystemProperties(),
-      ...this.data,
-    };
-  }
-
-  /**
    * @private
    *
    * @return {Object}
    */
   getSystemProperties() {
     const rawDocumentTransition = {
-      ...super.toJSON(),
       $id: this.getId(),
       $type: this.getType(),
       $entropy: this.getEntropy(),
@@ -174,6 +117,20 @@ class DocumentCreateTransition extends AbstractDocumentTransition {
     }
 
     return rawDocumentTransition;
+  }
+
+  /**
+   * Create document transition from JSON
+   *
+   * @param {RawDocumentCreateTransition} rawDocumentTransition
+   * @param {DataContract} dataContract
+   *
+   * @return {DocumentCreateTransition}
+   */
+  static fromJson(rawDocumentTransition, dataContract) {
+    return AbstractDocumentTransition.fromJson(
+      DocumentCreateTransition, rawDocumentTransition, dataContract,
+    );
   }
 }
