+++ lib/document/stateTransition/documentTransition/DocumentDeleteTransition.js
@@ -3,9 +3,10 @@ const AbstractDocumentTransition = require('./AbstractDocumentTransition');
 class DocumentDeleteTransition extends AbstractDocumentTransition {
   /**
    * @param {RawDocumentDeleteTransition} rawTransition
+   * @param {DataContract} dataContract
    */
-  constructor(rawTransition) {
-    super(rawTransition);
+  constructor(rawTransition, dataContract) {
+    super(rawTransition, dataContract);
 
     this.id = rawTransition.$id;
     this.type = rawTransition.$type;
@@ -39,13 +40,12 @@ class DocumentDeleteTransition extends AbstractDocumentTransition {
   }
 
   /**
-   * Get document transition as a plain object
+   * Get transition system properties
    *
-   * @return {RawDocumentDeleteTransition}
+   * @return {{ $id: string, $type: string }}
    */
-  toJSON() {
+  getSystemProperties() {
     return {
-      ...super.toJSON(),
       $id: this.getId(),
       $type: this.getType(),
     };
