+++ lib/document/stateTransition/documentTransition/DocumentReplaceTransition.js
@@ -1,6 +1,3 @@
-const lodashGet = require('lodash.get');
-const lodashSet = require('lodash.set');
-
 const AbstractDocumentTransition = require('./AbstractDocumentTransition');
 
 class DocumentReplaceTransition extends AbstractDocumentTransition {
@@ -9,9 +6,7 @@ class DocumentReplaceTransition extends AbstractDocumentTransition {
    * @param {DataContract} dataContract
    */
   constructor(rawTransition, dataContract) {
-    super(rawTransition);
-
-    this.dataContract = dataContract;
+    super(rawTransition, dataContract);
 
     const data = { ...rawTransition };
 
@@ -88,64 +83,12 @@ class DocumentReplaceTransition extends AbstractDocumentTransition {
   }
 
   /**
-   * Get document transition as a plain object
-   * with encoded binary fields
-   *
-   * @return {RawDocumentCreateTransition}
-   */
-  toJSON() {
-    const data = JSON.parse(
-      JSON.stringify(this.data),
-    );
-
-    const encodedProperties = this.dataContract.getEncodedProperties(
-      this.getType(),
-    );
-
-    Object.keys(encodedProperties)
-      .forEach((propertyPath) => {
-        const property = encodedProperties[propertyPath];
-
-        if (property.contentEncoding === 'binary') {
-          const value = lodashGet(data, propertyPath);
-          if (value !== undefined) {
-            lodashSet(
-              data,
-              propertyPath,
-              Buffer.from(value)
-                .toString('utf-8'),
-            );
-          }
-        }
-      });
-
-    return {
-      ...this.getSystemProperties(),
-      ...data,
-    };
-  }
-
-  /**
-   * Get document transition as a plain object
-   * without encoded binary fields
-   *
-   * @return {RawDocumentCreateTransition}
-   */
-  toObject() {
-    return {
-      ...this.getSystemProperties(),
-      ...this.data,
-    };
-  }
-
-  /**
    * @private
    *
    * @return {Object}
    */
   getSystemProperties() {
     const rawDocumentTransition = {
-      ...super.toJSON(),
       $id: this.getId(),
       $type: this.getType(),
       $revision: this.getRevision(),
@@ -156,6 +99,20 @@ class DocumentReplaceTransition extends AbstractDocumentTransition {
     }
 
     return rawDocumentTransition;
+  }
+
+  /**
+   * Create document transition from JSON
+   *
+   * @param {RawDocumentReplaceTransition} rawDocumentTransition
+   * @param {DataContract} dataContract
+   *
+   * @return {DocumentReplaceTransition}
+   */
+  static fromJson(rawDocumentTransition, dataContract) {
+    return AbstractDocumentTransition.fromJson(
+      DocumentReplaceTransition, rawDocumentTransition, dataContract,
+    );
   }
 }
